worker-service/requirements.txt
fastapi==0.115.0
google-cloud-firestore==2.19.0
uvicorn[standard]==0.30.6
